[refactoring] file name, package name 변경
